Make threshold configurable for TreeOverlay (#426)
